project/backend/requirements.txt
fastapi
uvicorn
prometheus_client
langgraph
langchain-ollama
langchain-mistralai
langchain-community
beautifulsoup4
chromadb
pytest-playwright
python-dotenv
GitPython
pytest-html
